rename zendesk -> zendesk support (#1318)
