+++ airbyte-config/init/src/main/resources/config/STANDARD_SOURCE_DEFINITION/d29764f8-80d7-4dd7-acbe-1a42005ee5aa.json
@@ -1,7 +1,7 @@
 {
   "sourceDefinitionId": "d29764f8-80d7-4dd7-acbe-1a42005ee5aa",
-  "name": "Zendesk",
-  "dockerRepository": "airbyte/source-zendesk-singer",
+  "name": "Zendesk Support",
+  "dockerRepository": "airbyte/source-zendesk-support-singer",
   "dockerImageTag": "0.1.0",
-  "documentationUrl": "https://hub.docker.com/r/airbyte/source-zendesk-singer"
+  "documentationUrl": "https://hub.docker.com/r/airbyte/source-zendesk-support-singer"
 }
