+++ airbyte-integrations/connectors/source-zendesk-support-singer/Dockerfile
@@ -2,9 +2,9 @@ FROM airbyte/integration-base-singer:dev
 
 RUN apt-get update && apt-get install -y bash && rm -rf /var/lib/apt/lists/*
 
-ENV CODE_PATH="source_zendesk_singer"
-ENV AIRBYTE_IMPL_MODULE="source_zendesk_singer"
-ENV AIRBYTE_IMPL_PATH="SourceZendeskSinger"
+ENV CODE_PATH="source_zendesk_support_singer"
+ENV AIRBYTE_IMPL_MODULE="source_zendesk_support_singer"
+ENV AIRBYTE_IMPL_PATH="SourceZendeskSupportSinger"
 
 WORKDIR /airbyte/integration_code
 COPY $CODE_PATH ./$CODE_PATH
@@ -12,4 +12,4 @@ COPY setup.py ./
 RUN pip install ".[main]"
 
 LABEL io.airbyte.version=0.1.0
-LABEL io.airbyte.name=airbyte/source-zendesk-singer
+LABEL io.airbyte.name=airbyte/source-zendesk-support-singer
