+++ airbyte-integrations/connectors/source-zendesk-support-singer/sample_files/full_catalog.json
