+++ docs/SUMMARY.md
@@ -36,7 +36,7 @@
     * [Slack](integrations/sources/slack.md)
     * [Stripe](integrations/sources/stripe.md)
     * [Twilio](integrations/sources/twilio.md)
-    * [Zendesk](integrations/sources/zendesk.md)
+    * [Zendesk Support](integrations/sources/zendesk-support.md)
     * [Zoom](integrations/sources/zoom.md)
   * [Destinations](integrations/destinations/README.md)
     * [BigQuery](integrations/destinations/bigquery.md)
